docker-image/src/requirements.txt
Flask==2.1.1
gunicorn==20.1.0
numpy==1.24.4
python-dotenv==1.0.1
django-cors-headers
psycopg2-binary>=2.9
firebase-admin==5.0.2
replicate==0.24.0
transformers==4.38.2
llama_index==0.10.19
faiss-cpu
langchain_community
langchain_openai
alembic>=1.12.0
SQLAlchemy>=1.4
Flask-JWT-Extended>=4.4.0
streamlit>=1.44.0
PyPDF2
textract
tiktoken